k40-socket-server/status/status_socket.py
import eventlet
import socketio

sio = socketio.Server()




@sio.on('connect')
def connect(sid, environ):
    print('connect ', sid)


@sio.on('disconnect')
def disconnect(sid):
    print('disconnect ', sid)



