k40-socket-server/requirements.txt
python-socketio==3.1.2
eventlet==0.24.1
rx==3.0.0a3